Added more visual feedback from UI
- While the dashboard is loading data in the background, each of the
performance meters states "Gathering Data...", and promptly changes once
data has been pulled.

- The deployment progress text now states "GameServer Deployed / Updated
Successfully!" at the end of a successful deployment or update of
existing gameserver.  This is a light side-step to the original green
popup notifying the user of the same information.  If the user cancels a
deployment, or starts another one, the text will revert back to
"Download / Installation Progress:".